% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwsp_model.R
\name{bwsp_model}
\alias{bwsp_model}
\title{Fit Bayesian model to time-to-event data}
\usage{
bwsp_model(datstan, chains = 4, iter = 11000, warmup = 1000)
}
\arguments{
\item{datstan}{named list of data for the stanmodel; output of \code{\link{tte2priordat}}}

\item{chains}{number of Markov chains to run}

\item{iter}{total number of iterations per chain (including warmup)}

\item{warmup}{number of warmup iterations per chain}
}
\value{
A list with components:
\itemize{
\item \code{fit}: fitted model object; a \code{stanfit} object for \code{"w"} and \code{"pgw"},
or a list of two \code{stanfit} objects (\code{$uncens}, \code{$cens}) for \code{"dw"}
\item \code{args_list}: list of model specifications, including the chosen
time-to-event distribution and prior settings
}
}
\description{
Fits a Bayesian model to time-to-event (tte) data
for the purpose of performing Weibull shape parameter signal detection tests
with \code{\link{bwsp_test}}.
}
\details{
The function applies the \code{\link[rstan]{sampling}} command with the No U-Turn sampler
to fit a Bayesian model to tte data.

The posterior is proportional to the likelihood times the prior.
The likelihood used in ML estimation is
\deqn{\mathcal{L}(t) = \prod_{i=1}^N S(t_i)^{1-d_i}\cdot f(t_i)^{d_i}}
with \eqn{S(t)}  being the survival function of the chosen distribution, \eqn{f(t)} the
density \insertCite{nikulin2016}{WSPsignal}, and \eqn{(t_i, d_i)}
the (right-censored) tte observations.

Prior and tte distribution are specified in the previous data preparation step
with function \code{\link{tte2priordat}}.
}
\examples{

head(tte)

# prep the data
# we formalize a prior belief (here "no association
# between drug and event", therefore prior mean = 1 for shape parameter)
# and reformat our tte data to fit the model in the following
dat_list = tte2priordat(dat = tte,   # reformat the data
                      tte.dist = "w",
                      prior.dist = "ll", 
                      scale.mean = 1, 
                      scale.sd = 10,
                      shape.mean = 1, 
                      shape.sd = 10)

# model fitting
mod = bwsp_model(datstan = dat_list,      # fit the model       
                 chains = 4,              
                 iter = 110,             # (posterior sample is
                 warmup = 10)            # small for demo purpose)
mod$fit


}
\references{
\insertAllCited{}
}
