% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwsp_test.R
\name{bwsp_test}
\alias{bwsp_test}
\title{Bayesian Weibull Shape Parameter Test}
\usage{
bwsp_test(
  mod.output,
  cred.level = 0.8,
  ci.type = "HDI",
  sensitivity.option = 2
)
}
\arguments{
\item{mod.output}{model output resulting from \code{\link{bwsp_model}}}

\item{cred.level}{numeric or vector of credibility levels; default is 0.8}

\item{ci.type}{character indicating whether to extract an equal tailed
interval (\code{"ETI"}) or highest posterior density interval (\code{"HDI"}) as
posterior credibility interval (CI) for the BWSP test; default is \code{"HDI"}}

\item{sensitivity.option}{numeric value out of \code{1,2,3}; combination rule to deduct a
binary outcome (signal/no signal) from one or two shape parameter tests;
default is \code{sensitivity.option = 2} (see details)}
}
\value{
binary vector, 0 if \eqn{H_0} is accepted (no signal), 1 if \eqn{H_1} is accepted (signal)
}
\description{
Bayesian Weibull Shape Parameter (BWSP) test of the constant hazard (null-)hypothesis,
based on the shape parameter(s) of Weibull family of distributions.
}
\section{Test concept}{


The BWSP test is based on the principle of (non-)constant hazard
\insertCite{cornelius2012}{WSPsignal}
which associates a constant hazard function with the absence of a drug-adverse event
association
and a non-constant hazard with the presence of a drug-adverse event
association.

This can be formalized as the following hypotheses
depending on the underlying model:

\tabular{lcc}{
\tab \eqn{H_0} \tab \eqn{H_1} \cr
hypothesis \tab constant hazard function \tab non-constant hazard function \cr
under Weibull model \tab \eqn{\nu = 1} \tab \eqn{\nu \neq 1} \cr
under double Weibull model \tab \eqn{\nu_1 = 1 \text{ and } \nu_2 = 1} \tab \eqn{\nu_1 \neq 1 \text{ or } \nu_2 \neq 1} \cr
under Power generalized Weibull model \tab \eqn{\nu = 1 \text{ and } \gamma = 1} \tab \eqn{\nu \neq 1 \text{ or } \gamma \neq 1} \cr
}
}

\section{Bayesian test components}{


Information on the Bayesian
variant of the Power Generalized Weibull (PGW) shape
parameter test can be found in \insertCite{dyck2024bpgwsppreprint;textual}{WSPsignal}.
The same concept applies to the construction of the Bayesian Weibull and double
Weibull shape parameter test.

The region of practical equivalence (ROPE) represents the expected parameter value under \eqn{H_0}.
The posterior credibility interval(s) (CI) represent the posterior distribution
of each shape parameter.
For the ROPE, the function sets up an equal-tailed interval (ETI)
\deqn{[q_{(1-\alpha)/2}, q_{(1+\alpha)/2}]}
based on the quantiles \eqn{q} of the shape parameters' prior distributions under
\eqn{H_0} at a chosen credibility level \eqn{1 - \alpha}.

For the posterior CI, the function calculates either an ETI at the same credibility level obtained from
the empirical quantiles of the posterior distribution per shape parameter
or a highest density interval (HDI, \insertCite{kruschke2015;textual}{WSPsignal})
\deqn{HDI(\nu) = \{\nu \; |\; p_1(\nu) \geq w\} \text{ with } w\in [0,1] 
\text{ such that} \int_{\nu \; | \; p_1(\nu) \geq w} p_1(\nu|t)\;  d\nu = 1 - \alpha}

where \eqn{\nu} is one of the shape parameters, \eqn{p_1} it's posterior density and \eqn{t} the time variable.

The CI+ROPE test \insertCite{kruschke2018}{WSPsignal} checks the
relationship between ROPE and posterior CI leading to either acceptance,
rejection or no decision regarding the null hypothesis for a single shape parameter.
Sensitivity options to generate a binary outcome, i.e. a signal or not, from CI+ROPE test results
based on one (in case of \code{"w"}) or two (in case of \code{"dw", "pgw"}) shape parameters are:
\tabular{ccccc}{
HDI+ROPE \tab HDI+ROPE \tab combination \tab combination \tab combination \cr
outcome \tab outcome \tab rule \tab rule \tab rule \cr
for shape_1\tab for shape_2 \tab (\code{sensitivity.option = 1}) \tab (\code{sensitivity.option = 2}) \tab (\code{sensitivity.option = 3}) \cr
rejection \tab (none) \tab signal \tab signal \tab signal \cr
acceptance \tab (none) \tab - \tab - \tab - \cr
no decision \tab (none) \tab signal \tab - \tab - \cr
rejection \tab rejection \tab signal \tab signal \tab signal \cr
acceptance \tab rejection \tab signal \tab - \tab - \cr
rejection \tab acceptance \tab signal \tab - \tab - \cr
acceptance \tab acceptance \tab - \tab - \tab - \cr
no decision \tab rejection \tab signal \tab signal \tab - \cr
no decision \tab acceptance \tab - \tab - \tab - \cr
rejection \tab no decision \tab signal \tab signal \tab - \cr
acceptance \tab no decision \tab - \tab - \tab - \cr
no decision \tab no decision \tab signal \tab - \tab - \cr
}

The hypotheses as stated above (see test concept) are implemented in \code{sensitivity.option = 1} whereas
\code{sensitivity.option = 2} and \code{sensitivity.option = 3} lead to a signal in fewer cases.

More details on the CI+ROPE test, recommendations for interval specifications
and the combination rules
can be found in \insertCite{kruschke2018;textual}{WSPsignal} and
\insertCite{dyck2024bpgwsppreprint;textual}{WSPsignal}.
}

\examples{
#### Exemplary conduction of a test from data and prior to test result:

# under weibull model:

# 1. prior specification
# we formalize a prior belief (here "no association
# between drug and event", therefore prior mean = 1 for shape parameter)
# and reformat our tte data to fit the model in the following
dat_list = tte2priordat(dat = tte,   # reformat the data
                      tte.dist = "w",
                      prior.dist = "ll", 
                      scale.mean = 1, 
                      scale.sd = 10,
                      shape.mean = 1, 
                      shape.sd = 10)

# 2. model fitting
fit = bwsp_model(datstan = dat_list,      # fit the model       
                 chains = 4,              
                 iter = 110,             # (posterior sample is
                 warmup = 10)            # small for demo purpose)
fit$fit

# 3. BWSP test
bwsp_test(mod.output = fit,
          cred.level = 0.8,
          ci.type = "HDI",
          sensitivity.option = 2)

# under pgw model:

# 1. prior specification
# we formalize a prior belief (here "no association
# between drug and event", therefore prior mean = 1 for both shape parameters)
# and reformat our tte data to fit the model in the following
dat_list = tte2priordat(dat = tte,          # reformat the data
                      tte.dist = "pgw",
                      prior.dist = "ll",
                      scale.mean = 1, 
                      scale.sd = 10,
                      shape.mean = 1, 
                      shape.sd = 10,
                      powershape.mean = 1, 
                      powershape.sd = 10)

# 2. model fitting
fit = bwsp_model(datstan = dat_list,     # fit the model      
                  chains = 4,              
                  iter = 110,            # (posterior sample
                  warmup = 10)           # is small for demo purpose)

# 3. BWSP test
bwsp_test(mod.output = fit,
          cred.level = 0.8,
          ci.type = "HDI", 
          sensitivity.option = 2)



}
\references{
\insertAllCited{}
}
