% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.non_conv_cases.R
\name{eval.non_conv_cases}
\alias{eval.non_conv_cases}
\title{Evaluate number of non-convergence cases by tte and prior distribution}
\usage{
eval.non_conv_cases(pc_list)
}
\arguments{
\item{pc_list}{list of simulation parameters generated with \code{\link{sim.setup_sim_pars}}}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{tte.dist}: The tte distribution as grouping factor.
\item \code{prior.dist}: The prior distribution as grouping factor.
\item \code{total.planned}: The total number of planned repetitions in this group.
\item \code{total.notrun}: The total number of repetitions that were not run in this group.
\item \code{prop.notrun}: The proportion of repetitions that were not run in this group.
}
}
\description{
Only for Bayesian estimation approach.
Summarizes the number of planned vs. not successfully run simulations grouped by \code{tte.dist}
and \code{prior.dist}. The purpose of the table is to guide the
time-to-event (tte) and prior distributional choices (along with other diagnostics
such as \code{\link{eval.execution_times}} and \code{\link{eval.eff_sample_sizes}}).
}
\details{
Calculations are based on the stored result batch files of the simulation study.
Unsuccessfully run simulations are missing (supposedly due to non-convergence during
model estimation).
}
\seealso{
\code{\link{eval.execution_times}}, \code{\link{eval.eff_sample_sizes}}
}
