% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.rank_auc.R
\name{eval.rank_auc}
\alias{eval.rank_auc}
\title{Ranking of WSP test configurations}
\usage{
eval.rank_auc(
  perf,
  test.type.subset = c("bwsp", "fwsp"),
  tte.dist.subset = c("w", "dw", "pgw"),
  prior.dist.subset = c("fg", "fl", "gg", "ll"),
  prior.sd.subset = NULL
)
}
\arguments{
\item{perf}{data frame containing performance results for WSP tests returned
by \code{\link{eval.calc_perf}}}

\item{test.type.subset}{character to filter for Bayesian and frequentist
WSP test types to be considered in the ranking;
must be a subset of \code{c("bwsp", "fwsp")}}

\item{tte.dist.subset}{character to filter for the time-to-event (tte)
distributions considered in the ranking, must be a subset of
\code{c("w", "dw", "pgw")}}

\item{prior.dist.subset}{character to filter for the prior distribution
(relevant only for BWSP tests), must be a subset of
\code{c("fg", "fl", "gg", "ll")}}

\item{prior.sd.subset}{numeric to filter for the prior standard deviation
(relevant only for BWSP tests), must be a subset of prior.sds considered
in the simulation}
}
\value{
A list containing
\itemize{
\item \code{$rank.tab}: Ranking of fit and WSP test specifications according
to AUC averaged over all sample scenarios (for BWSP tests given a correct
specification of prior belief)
\item \code{$effect.of.N}: Effect of sample size on AUC for the optimal fit
and WSP test (for BWSP tests given a correct specification of prior belief)
\item \code{$effect.of.br}: Effect of background rate on AUC for the optimal fit
and WSP test (for BWSP tests given a correct specification of prior belief)
\item \code{$effect.of.adr.rate}: Effect of ADR rate on AUC for the optimal fit
and WSP test (for BWSP tests given a correct specification of prior belief)
\item \code{$effect.of.adr.when}: Effect of true expected event times on AUC
for the optimal fit and WSP test (for BWSP tests given a correct
specification of prior belief)
\item \code{$effect.of.adr.relsd}: Effect of relative standard deviation of event
time on AUC for the optimal fit and WSP test (for BWSP tests given a correct
specification of prior belief)
\item \code{$effect.of.dist.prior.to.truth}: Effect of distance of prior belief
to true \code{adr.when} on AUC for the optimal fit and WSP test (only BWSP)
}
}
\description{
Ranks all model and test specifications grouped by simulation scenarios in terms
of the corresponding area under the curve (AUC) value for Weibull Shape Parameter
(WSP) tests.
}
\details{
For definitions of the performance metrics \code{AUC, FPR, TPR, FNR}
and \code{TNR} returned in output,
see the details section of \code{\link{eval.calc_perf}}.

The filter mechanism enables filtering for a subset of test specifications.
This is helpful for example when tte distributions, prior distributions or
an estimation approach are no longer under consideration for instance after
inspecting the model diagnostics with \code{\link{eval.execution_times}},
\code{\link{eval.non_conv_cases}} and \code{\link{eval.eff_sample_sizes}}.
}
\examples{
\dontrun{
# loading of performance metrics returned by eval.calc_perf function,
load(paste0(pc_list$add$resultpath, "/perf.RData"))

# ranking of all WSP tests considered in simulation setup
rank = eval.rank_auc(perf)

# ranking of subset of all Bayesian WSP tests considered in the simulation setup
rank_b = eval.rank_auc(perf, test.type.subset = "bwsp")

# ranking of subset of all frequentist Weibull tests considered
# in the simulation setup
rank_w = eval.rank_auc(perf, tte.dist.subset = "w")

# ranking of subset of all Bayesian WSP tests with "ll"-prior considered in 
# the simulation setup
rank_b_ll = eval.rank_auc(perf, test.type.subset = "bwsp",
                       prior.dist.subset = "ll")
                       
}

}
