% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.roc_curve.R
\name{eval.roc_curve}
\alias{eval.roc_curve}
\title{Plot ROC curves for top WSP test specifications}
\usage{
eval.roc_curve(rank.tab, n = 10)
}
\arguments{
\item{rank.tab}{data frame of ranked test specifications obtained from
\code{\link{eval.rank_auc}} (\code{output$rank.tab})}

\item{n}{number of top-ranked test specifications to plot (10 by default)}
}
\value{
A ggplot object displaying ROC curves with shaded AUC regions.
}
\description{
Plots receiver operating characteristic (ROC) curves for the top-ranked test
specifications.
}
\details{
The function returns the receiver ROC curves for the top n WSP test
configurations based
on the ranking returned by \code{\link{eval.rank_auc}} by plotting the
true positive rate (TPR) on the y-axis against the false positive rate (FPR)
on the x-axis \insertCite{fawcett2004}{WSPsignal}.
Here, we use the ROC curve with one threshold based on equal numbers
of ADR-positive and control scenarios.

For definitions of the performance metrics \code{AUC, FPR, TPR, FNR}
and \code{TNR} returned in printed output,
see the details section of \code{\link{eval.calc_perf}}.

The ggplot output can be adjusted to individual needs by adding \code{ggplot2}
layers to the output.
}
\examples{
\dontrun{
# Given object rank as output of eval.rank_auc:
eval.roc_curve(rank$rank.tab, n = 1) # top 1 test
eval.roc_curve(rank$rank.tab, n = 10) # top 10 tests

# exemplary further processing of the plot
library(patchwork) # above plots next to each other
# Left plot: n=1, no legend, no title
p1 <- eval.roc_curve(rank$rank.tab, n = 1) +
  ggplot2::guides(color = "none", fill = "none") +
  ggplot2::labs(title = NULL)
# Right plot: n=10, legend, no title
p2 <- eval.roc_curve(rank$rank.tab, n = 10) +
  ggplot2::labs(title = NULL)
# Combine with patchwork and add global title
(p1 | p2) + patchwork::plot_annotation(title = "ROC curves")
}

}
\references{
\insertAllCited{}
}
