% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwsp_test.R
\name{fwsp_test}
\alias{fwsp_test}
\title{Frequentist Shape Parameter Test}
\usage{
fwsp_test(mod.output, cred.level = 0.96)
}
\arguments{
\item{mod.output}{model output resulting from \code{\link{fwsp_model}}}

\item{cred.level}{numeric or vector of confidence levels (i.e. 1 - significance level)
for the test(s) to be performed}
}
\value{
binary vector, 0 if \eqn{H_0} is accepted, 1 if \eqn{H_1} is rejected
}
\description{
Frequentist hypothesis test based on the shape parameter(s) of a distribution of the
Weibull family.
}
\details{
This function tests the null hypothesis that the shape parameter(s) of the
Weibull family distribution are equal to one.
The distribution specific definition of the null and alternative hypotheses
can be seen in \insertCite{sauzet2022;textual}{WSPsignal}.

For the \code{"w"} and \code{"dw"} case, the model output is a summary of a
\code{survival::Survreg} outcome which provides
\eqn{ln(1 / \nu)} as transform of the shape parameter estimate \eqn{\nu}.
The transform \eqn{ln(1 / \nu) = 0} under the null hypothesis \eqn{\nu = 1}.
The shape parameter test is performed on the transform
equivalent to performing the test based on the shape parameter itself.

For the \code{"pgw"} case, the shape parameter test is performed on the logarithmized
parameter estimates, i.e.
\code{fwsp_test} tests the null hypothesis that the logarithm of the shape parameters
of the power generalized Weibull distribution are equal to zero based on the
shape estimates and their estimated standard errors extracted from the estimated
Hessian matrix.
Issues with standard error calculation from the estimated Hessian matrix
may lead to a \code{NA} test results which are then transformed to no signal (0)
following \insertCite{sauzet2022;textual}{WSPsignal}.
}
\examples{
# fit a model
mod = fwsp_model(dat = tte)
mod
# perform the shape parameter test at credibility level 0.95 
# or significance level 0.05
fwsp_test(mod.output = mod, cred.level = 0.95)

}
\references{
\insertAllCited{}
}
