% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgw_functions.R
\name{pgw}
\alias{pgw}
\alias{spgw}
\alias{hpgw}
\alias{ppgw}
\alias{dpgw}
\alias{qpgw}
\alias{rpgw}
\title{The power generalized Weibull distribution}
\usage{
spgw(x, scale = 1, shape = 1, powershape = 1, log = FALSE)

hpgw(x, scale = 1, shape = 1, powershape = 1, log = FALSE)

ppgw(x, scale = 1, shape = 1, powershape = 1)

dpgw(x, scale = 1, shape = 1, powershape = 1, log = FALSE)

qpgw(p, scale = 1, shape = 1, powershape = 1)

rpgw(n, scale = 1, shape = 1, powershape = 1)
}
\arguments{
\item{x}{vector of quantiles}

\item{scale}{scale parameter}

\item{shape}{shape parameter}

\item{powershape}{power shape parameter}

\item{log}{FALSE (default); if TRUE, the logarithm of the survival probability is returned}

\item{p}{vector of probabilities}

\item{n}{number of observations}
}
\value{
A numeric vector of

\itemize{
\item \code{spgw}: survival probabilities \eqn{S(x)} (or log-survival if \code{log = TRUE})
\item \code{hpgw}: hazard values \eqn{h(x)} (or log-hazard if \code{log = TRUE})
\item \code{ppgw}: cumulative distribution values \eqn{F(x)}
\item \code{dpgw}: density values \eqn{f(x)} (or log-density if \code{log = TRUE})
\item \code{qpgw}: quantiles corresponding to probabilities \code{p}
\item \code{rpgw}: random samples
}
}
\description{
Survival, hazard, cumulative distribution,
density, quantile and sampling function for the power generalized
Weibull (PGW) distribution with parameters \code{scale}, \code{shape} and \code{powershape}.
}
\details{
The survival function of the PGW distribution is
\deqn{
    S(x) = \exp \left\{ 1 - \left[ 1 + \left(\frac{x}{\theta}\right)^{\nu}\right]^{\frac{1}{\gamma}} \right\}.
}
with scale \eqn{\theta}, shape \eqn{\nu} and power shape parameter \eqn{\gamma}.

The hazard function is
\deqn{
    h(x) = \frac{\nu}{\gamma\theta^{\nu}}\cdot x^{\nu-1}\cdot  \left[ 1 + \left(\frac{x}{\theta}\right)^{\nu}\right]^{\frac{1}{\gamma-1}}
}
The cumulative distribution function is then \eqn{F(x) = 1 - S(x)} and the density function
is \eqn{f(x) = S(x)\cdot h(x)}. The quantile function is the inverse of the cumulative
distribution function \eqn{F^{-1}(x)}.

If both shape parameters equal 1, the PGW distribution reduces to the exponential distribution
(see \code{\link[stats]{dexp}}) with \eqn{\texttt{rate} = 1/\texttt{scale}}
If the power shape parameter equals 1, the PGW distribution reduces to the Weibull distribution
(see \code{\link[stats]{dweibull}}) with the same parametrization.

If parameter values are not specified, they are set as
\code{scale = 1, shape = 1, powershape = 1} per default.
}
\references{
\insertRef{nikulin2016}{WSPsignal}
}
