% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.datagen_tte.R
\name{sim.datagen_tte}
\alias{sim.datagen_tte}
\title{Generate simulated time-to-event data}
\usage{
sim.datagen_tte(genpar)
}
\arguments{
\item{genpar}{A vector containing 6 numeric elements:
\enumerate{
\item integer sample size \eqn{N},
\item background rate \eqn{br} in \eqn{(0,1]} (observed in population on average),
\item adverse drug reaction (ADR) rate \eqn{adr} in \eqn{[0,1]} as proportion of the background rate,
\item relative proportion of ADR mean time \eqn{m.rel} in \eqn{(0,1)} of the observation period (OP),
\item relative standard deviation \eqn{rel.sd} in \eqn{(0,1)} of the OP,
\item length of the OP \eqn{censor}.
}}
}
\value{
A data frame of size \eqn{N} with variables \code{time} (integer) indicating
event- or censoring time and \code{status} (binary) indicating whether the event was observed or
the observation was censored.
}
\description{
Simulation of time-to-event (tte) data.
}
\details{
After specification of the input, the data simulation works as follows:

The absolute number of events due to background causes (other than ADR) is
generated by a binomial distribution with probability \eqn{br}.
The absolute number of events caused by the ADR is generated with a
binomial distribution with probability \eqn{br\cdot adr} giving an expected
number of events within the data set is \eqn{n \cdot br (1 + adr)}.

For the \eqn{br} cases, the event-times are generated using a uniform distribution on
the interval \eqn{[0, censor]}. For the ADR cases, event-times are obtained from a normal
distribution. The mean of the normal distribution is specified as relative proportion
\eqn{m.rel} of the OP (e.g. 0.5 for the middle of the OP).
The standard deviation is defined as \eqn{ rel.sd \cdot censor}.
All generated event-times \eqn{\leq 0} (due to the normal distribution's support) are set to 1.
All generated event-times \eqn{\geq censor} are censored retroactively.
The continuous values are rounded to integer.
The data set is filled up with censored observations (status = 0) at time \eqn{censor}.

For more details, see \insertCite{dyck2024bpgwsppreprint;textual}{WSPsignal}.
}
\examples{
sim.datagen_tte(c(100, 0.1, 1, 0.5, 0.05, 365))

}
\references{
\insertAllCited{}
}
\seealso{
\link{tte}
}
