% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.run.R
\name{sim.run}
\alias{sim.run}
\title{Run simulation}
\usage{
sim.run(pc_list, subset_ind = NULL)
}
\arguments{
\item{pc_list}{list of parameter combinations generated with \code{\link{sim.setup_sim_pars}}}

\item{subset_ind}{vector of integers specifying which rows of \code{pc_list$pc_table}
to be considered in simulation runs; defaults to all rows}
}
\value{
No return value. This function is called for
running simulations and writing results to disk.
}
\description{
Runs simulations for all specified data-generating processes, model
and test configurations defined in \code{pc_list}. If simulations for some scenarios
have already been completed, the function resumes from the remaining scenarios.
}
\details{
This function executes simulations sequentially. For parallel
execution, see \code{\link{sim.run_parallel}}.
}
\examples{
\dontrun{
# Running simulation studies can take considerable time, especially when
# Bayesian estimation is included. Therefore, this example is not run.

# See ?sim.setup_sim_pars for an example creating pc_list.

# run all simulation scenarios
sim.run(pc_list) # run all simulation scenarios
 
# to run only a subset of simulation scenarios, specify row indices from
# pc_list$pc_table, e.g. the first 10 scenarios:
sim.run(pc_list, subset_ind = 1:10)
}


}
