% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.run_parallel.R
\name{sim.run_parallel}
\alias{sim.run_parallel}
\title{Run simulation (in parallel)}
\usage{
sim.run_parallel(pc_list, subset_ind = NULL)
}
\arguments{
\item{pc_list}{list of parameter combinations generated with \code{\link{sim.setup_sim_pars}}}

\item{subset_ind}{vector of integers specifying which rows of \code{pc_list$pc_table}
to be considered in simulation runs; defaults to all rows}
}
\value{
No return value. This function is called for
running simulations and writing results to disk.
}
\description{
Runs simulations in parallel for all specified data-generating processes, model
and test configurations defined in \code{pc_list}. If simulations for some scenarios
have already been completed, the function resumes from the remaining scenarios.
}
\details{
Parallelization needs to be initialized using the \code{\link[future]{plan}}
command (see example).
}
\examples{
\dontrun{
 # install.packages(future)
 future::plan(multisession, workers = future::availableCores()) # or another plan strategy
 sim.run_parallel(pc_list) # run all simulations
 
 
 # To run only a subset of simulation scenarios, specify the indices of the rows
 # in pc_list$pc_table that you want to run. For example, to run the first 10 scenarios:
 
 # install.packages(future)
 future::plan(multisession, workers = availableCores()) # or another plan strategy
 sim.run_parallel(pc_list, subset_ind = 1:10) # run first 10 simulation scenarios
}




}
