% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwsp_model.R
\name{bwsp_model}
\alias{bwsp_model}
\title{Fit Bayesian model to time-to-event data}
\usage{
bwsp_model(datstan, chains = 4, iter = 11000, warmup = 1000)
}
\arguments{
\item{datstan}{named list of data for the stanmodel; tte data can be
formated with \code{\link{tte2priordat}}}

\item{chains}{number of Markov chains to run}

\item{iter}{total number of iterations per chain (including warmup)}

\item{warmup}{number of warmup iterations per chain}

\item{tte.dist}{character indicating the modelling approach; options are
\code{"w", "dw", "pgw"} (see details)}

\item{prior.dist}{character indicating the prior distribution for the
parameters of the tte distribution; options are
\code{"fg", "fl", "gg", "ll"} (see details)}
}
\value{
A stanfit object or, in case of \code{tte.dist = "dw"}, a list of two stanfit objects
with \code{$uncens} containing the stanfit object
obtained from the model to the tte data as is, and \code{$cens} containing the stanfit object
obtained from the model to the tte data censored at mid of observation period.
}
\description{
Fits a Bayesian model to time-to-event (tte) data
for the purpose of performing Weibull shape parameter signal detection tests
with \code{\link{bwsp_test}}.
}
\details{
The function applies the \code{\link[rstan]{sampling}} command with the No U-Turn sampler
to fit a Bayesian model to
time-to-event data.

The posterior is proportional to the likelihood times the prior. The likelihood is
\deqn{\mathcal{L}(t| \Theta) = \prod_{i=1}^N S(t_i)^{1-d_i}\cdot f(t_i)^{d_i}}
with \eqn{S(t)} the survival function of the chosen distribution and \eqn{f(t)} the
density function \insertCite{nikulin2016}{WSPsignal}. The pair \eqn{(t_i, d_i)}
are the tte observations.

Prior and tte distribution are specified in data preparation with function
\code{\link{tte2priordat}}.
}
\examples{

head(tte)

# prep the data
# we formalize a prior belief (here "no association
# between drug and event", therefore prior mean = 1 for shape parameter)
# and reformat our tte data to fit the model in the following
dat_list = tte2priordat(dat = tte,   # reformat the data
                      tte.dist = "w",
                      prior.dist = "ll", 
                      scale.mean = 1, 
                      scale.sd = 10,
                      shape.mean = 1, 
                      shape.sd = 10)

# model fitting
fit = bwsp_model(datstan = dat_list,      # fit the model       
                 chains = 4,              
                 iter = 110,             # (posterior sample is
                 warmup = 10)            # small for demo purpose)
fit


}
\references{
\insertAllCited{}
}
