% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.eff_sample_sizes.R
\name{eval.eff_sample_sizes}
\alias{eval.eff_sample_sizes}
\title{Evaluate effective sample sizes}
\usage{
eval.eff_sample_sizes(
  pc_list,
  group.by = c("tte.dist", "prior.dist", "prior.sd"),
  threshold = 10000,
  verbose = TRUE
)
}
\arguments{
\item{pc_list}{list of simulation parameters generated with \code{\link{sim.setup_sim_pars}}}

\item{group.by}{character vector specifying grouping variables; must be a
subset of \code{c("tte.dist", "prior.dist", "prior.sd")}}

\item{threshold}{numeric threshold for effective sample size acceptable for
HDI+ROPE testing (10000 by default as recommended by \insertCite{kruschke2015;textual}{WSPsignal})}

\item{verbose}{logical; if \code{TRUE} (default), summary statistics
and plot are printed to the console}
}
\value{
A list with summary statistics (\verb{$summary}), a ggplot2 object (\verb{$plot}),
and the data (\verb{$df}) on which summary and plot are based.
}
\description{
Only for Bayesian estimation approach.
Summarizes effective sample sizes of the stan models fitted optionally
grouped by one or multiple model specifications.
This helps assess which of the modelling choices is suitable
for CI+ROPE testing (along with other diagnostics such as
\code{\link{eval.non_conv_cases}} and \code{\link{eval.execution_times}}).
}
\examples{
\dontrun{
# summarize effective sample sizes by tte and prior distribution and prior sd
eval.eff_sample_sizes(pc_list)

# summarize effective sample sizes only by tte distribution
eval.eff_sample_sizes(pc_list, group.by = "tte.dist")
}

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{eval.non_conv_cases}}, \code{\link{eval.execution_times}}
}
