% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.eff_sample_sizes.R
\name{eval.eff_sample_sizes}
\alias{eval.eff_sample_sizes}
\title{Evaluate effective sample sizes by tte and prior distribution}
\usage{
eval.eff_sample_sizes(
  pc_list,
  group.by = c("tte.dist", "prior.dist", "prior.sd"),
  threshold = 10000
)
}
\arguments{
\item{pc_list}{a list containing simulation parameters (see \code{\link{sim.setup_sim_pars}})}

\item{threshold}{numeric threshold for effective sample size acceptable for
HDI+ROPE testing (10000 by default as recommended by \insertCite{kruschke2015;textual}{WSPsignal})}
}
\value{
A list with summary statistics (\verb{$summary}), a ggplot2 object (\verb{$plot}),
and the data (\verb{$df}) on which summary and plot are based.
}
\description{
Only for Bayesian estimation approach.
Summarizes and visualizes effective sample sizes of the stan models fitted during
the simulation study, grouped by time-to-event (tte) and prior distribution types.
This helps assess which of the tte and prior distribution choices is suitable
for HDI+ROPE testing (along with other diagnostics such as
\code{\link{eval.non_conv_cases}} and \code{\link{eval.execution_times}}).
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{eval.non_conv_cases}}, \code{\link{eval.execution_times}}
}
