% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.execution_times.R
\name{eval.execution_times}
\alias{eval.execution_times}
\title{Evaluate execution times by tte and prior distribution}
\usage{
eval.execution_times(
  pc_list,
  group.by = c("tte.dist", "prior.dist", "prior.sd")
)
}
\arguments{
\item{pc_list}{list of simulation parameters generated with \code{\link{sim.setup_sim_pars}}}
}
\value{
A list with summary statistics (\verb{$summary}), a ggplot2 object (\verb{$plot}),
and the data (\verb{$df}) on which summary and plot are based.
}
\description{
Only for Bayesian estimation approach.
Summarizes and visualizes execution times of the models fitted during the simulation study grouped
by time-to-event (tte) and prior distribution types to guide the
tte and prior distributional choices (along with other diagnostics such as
\code{\link{eval.non_conv_cases}} and \code{\link{eval.eff_sample_sizes}}).
}
\seealso{
\code{\link{eval.non_conv_cases}}, \code{\link{eval.eff_sample_sizes}}
}
