% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwsp_model.R
\name{fwsp_model}
\alias{fwsp_model}
\title{Fit frequentist model to time-to-event data}
\usage{
fwsp_model(dat, tte.dist = c("dw"))
}
\arguments{
\item{dat}{data frame or matrix with time information in first column and
event information (binary status) in second column}

\item{tte.dist}{character specifying the distribution for the
model out of \code{"w", "dw", "pgw"} (see details)}
}
\value{
A list with components depending on \code{tte.dist}:
\itemize{
\item \code{fit}: fitted model object
(\code{summary.survreg} for \code{"w"}, a list of two \code{summary.survreg} objects for \code{"dw"},
and an \code{nlm} output list for \code{"pgw"})
\item \code{estimates}: data frame of parameter estimates in standard parametrization (scale, shape, powershape)
\item \code{tte.dist}: character indicating the fitted tte distribution
}
}
\description{
Fits a frequentist model to time-to-event (tte) data via maximum
likelihood (ML) estimation.
}
\details{
The model can be a Weibull \code{("w")},
a double Weibull (\code{"dw"}, estimating two Weibull models - one to the data as is and
one to the data censored at mid of observation period),
or a power generalized Weibull \code{("pgw")} model.

The likelihood used in ML estimation is
\deqn{\mathcal{L}(t) = \prod_{i=1}^N S(t_i)^{1-d_i}\cdot f(t_i)^{d_i}}
with \eqn{S(t)}  being the survival function of the chosen distribution, \eqn{f(t)} the
density \insertCite{nikulin2016}{WSPsignal}, and \eqn{(t_i, d_i)}
the (right-censored) tte observations.

For the estimation of the Weibull models\code{("w", "dw")}, the \code{\link[survival]{survreg}}
function (with no covariates) is used.

The \code{"pgw"} model is estimated by numerically minimizing the corresponding
negative log-likelihood function with \code{\link[stats]{nlm}}.
}
\examples{
head(tte)
fwsp_model(tte, tte.dist = "w") # Weibull model
fwsp_model(tte, tte.dist = "dw") # double Weibull model
fwsp_model(tte, tte.dist = "pgw") # power generalized Weibull model


}
\references{
\insertAllCited{}
}
