% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscu.R
\docType{data}
\name{muscu}
\alias{muscu}
\title{Simulated musculoskeletal pain time-to-event dataset}
\format{
A data frame with 19 777 rows and 2 variables:
\describe{
\item{time}{event time (in days) or censoring time (365 days),}
\item{status}{event indicator; 1 = event observed, 0 = censored.}
}
}
\usage{
muscu
}
\description{
A simulated time-to-event (tte) dataset representing musculoskeletal pain
following bisphosphonate intake.
}
\details{
The data was generated using
\code{\link{sim.datagen_tte(genpar = c(19777, 0.01, 0.89, 160/365, 0.1, 365))}}
with parameters derived from the case study presented in
\insertCite{dyck2024bpgwsppreprint;textual}{WSPsignal}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{sim.datagen_tte}}
}
\keyword{datasets}
