% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.merge_results.R
\name{sim.merge_results}
\alias{sim.merge_results}
\title{Merge result table batches from simulation study}
\usage{
sim.merge_results(pc_list, save = TRUE)
}
\arguments{
\item{pc_list}{list of parameter combinations generated with \code{\link{sim.setup_sim_pars}}}

\item{save}{if \code{TRUE} (default), merged table is saved as res_b.RData or
res_f.RData, respectively, in same path where batches are stored;
else, result table(s) is/are returned to global environment}
}
\value{
Dataframe or list of two dataframes containing all simulation results (one repetition of one
simulation scenario per row).
}
\description{
Merges result table batches from simulation study obtained from using
\code{\link{sim.run}} or \code{\link{sim.run_parallel}} as preparation for evaluation..
}
