% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.run.R
\name{sim.run}
\alias{sim.run}
\title{Run simulation}
\usage{
sim.run(pc_list, subset_ind = NULL)
}
\arguments{
\item{pc_list}{list of parameter combinations generated with \code{\link{sim.setup_sim_pars}}}

\item{subset_ind}{vector of integers specifying which rows of \code{pc_list$pc_table}
to be considered in simulation runs; defaults to all rows}
}
\description{
Runs simulations for all specified data-generating processes, model
and test configurations defined in \code{pc_list}. If simulations for some scenarios
have already been completed, the function resumes from the remaining scenarios.
}
\details{
This function executes simulations sequentially. For parallel
execution, see \code{\link{sim.run_parallel}}.
}
\examples{
 \dontrun{
 sim.run(pc_list) # run all simulations
 
 
 # To run only a subset of simulation scenarios, specify the indices of the rows
 # in pc_list$pc_table that you want to run. For example, to run the first 10 scenarios:
 
 sim.run(pc_list, subset_ind = 1:10) # run first 10 simulation scenarios
}


}
