% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.run_parallel.R
\name{sim.run_parallel}
\alias{sim.run_parallel}
\title{Run simulation (in parallel)}
\usage{
sim.run_parallel(pc_list, subset_ind = NULL)
}
\arguments{
\item{pc_list}{list of parameter combinations generated with \code{\link{sim.setup_sim_pars}}}

\item{subset_ind}{vector of integers specifying which rows of \code{pc_list$pc_table}
to be considered in simulation runs; defaults to all rows}
}
\value{
No return value. This function is called for
running simulations and writing results to disk.
}
\description{
Runs simulations in parallel for all specified data-generating processes, model
and test configurations defined in \code{pc_list}. If simulations for some scenarios
have already been completed, the function resumes from the remaining scenarios.
}
\details{
Parallelization needs to be initialized using the \code{\link[future]{plan}}
command (see example).
}
\examples{
\dontrun{
# Running simulation studies can take considerable time, especially when
# Bayesian estimation is included. Therefore, this example is not run.

# See ?sim.setup_sim_pars for an example creating pc_list.

# prep parallelization
# install.packages("future")
future::plan(future::multisession,
             workers = future::availableCores())

# run all simulation scenarios
sim.run_parallel(pc_list) 
 
# to run only a subset of simulation scenarios, specify row indices from
# pc_list$pc_table, e.g. the first 10 scenarios:
sim.run_parallel(pc_list, subset_ind = 1:10)
}




}
